refac: moving stripe to the gateways module
