+++ app/lib/payment/gateways/stripe/base.rb
@@ -1,0 +1,15 @@
+module Payment
+  module Gateways
+    module Stripe
+      class Base
+        ALLOWED_PAYMENT_METHODS = {
+          card: 'card'
+        }.freeze
+
+        def initialize
+          ::Stripe.api_key = RibonCoreApi.config[:stripe][:secret_key]
+        end
+      end
+    end
+  end
+end
